Добавит слайс extendedOfferSlice
