+++ src/store/selectors.ts
@@ -2,13 +2,8 @@ import { createSelector } from '@reduxjs/toolkit';
 import { State } from '../types/state';
 import { ReviewsListType } from '../types/review';
 import { NEARBY_OFFERS_LIMITED } from '../const';
-import { OfferType, ShortOfferListType } from '../types/offers';
+import { ShortOfferListType } from '../types/offers';
 
-const selectIsExtendedOfferLoading = (state: State): boolean =>
-  state.isExtendedOfferLoading;
-
-const selectExtendedOffer = (state: State): OfferType | null =>
-  state.extendedOffer;
 
 const selectIsNearbyOffersLoading = (state: State): boolean =>
   state.isNearbyOffersLoading;
@@ -32,8 +27,6 @@ const selectisSubmitReviewLoading = (state: State): boolean =>
 const selectErrorMessage = (state: State): string | null => state.error;
 
 export {
-  selectExtendedOffer,
-  selectIsExtendedOfferLoading,
   selectIsNearbyOffersLoading,
   selectIsReviewsListLoading,
   selectReviewsList,
