+++ src/types/state.ts
@@ -1,9 +1,10 @@
 import { store } from '../store';
 import { CityKeys } from './cities';
-import { OfferType, ShortOfferListType } from './offers';
+import { OfferType } from './offer';
+import { ShortOfferListType } from './offers';
 import { ReviewsListType } from './review';
 import { SortTypeKeys } from './sort';
-import { AuthorizationStatusKeys, UserInfo } from './user';
+import { AuthorizationStatusKeys, UserInfo} from './user';
 
 type InitialState = {
   currentCity: CityKeys;
